Use a size_t for ParseTree::init.

+++ include/parsetree.h
@@ -55,7 +55,7 @@ public:
   // ensures that the vector will never resize on us and invalidate our
   // ParseNode pointers.
   //
-  void init(uint32_t len);
+  void init(size_t len);
 
   bool operator==(const ParseTree& other) const {
     return !Root ? !other.Root : (other.Root ? *Root == *other.Root : false);
